Chapter 2
chap2 folder contains "real" implementation of lexer, but there are some errors with CM, so I want to tets it on simplified version of lexer, that is in folder chap22.

Errors fixed so far: 
- in `sources.cm` change smlnj-lib.cm to $/smlnj-lib.cm

Actually, ml-lex could produce *.lex.sml files successfully, but CM could not wire all dependencies listed in `sources.cm`.